Library Assistant Notebook Filtering
The example notebooks are published with results and images, which prevent efficient conversion to plain text for the library assistant and script to generate knowledge bases for generic LLM chatbot use.  Additional logic was added to filter notebook results, leaving dataframe headers in place and truncating the remainder of the dataframe for brevity, as well as removing images.

Also library assistant was moved and paths corrected.