Move hdf to submodules (#8)
* Reorganize RasDss into submodule

This functionality may need to expand, and is outside of core functions of the library so a submodule is more appropriate.  API should remain the same except for slight change in the import style.

* Move HDF to submodules

* move hdf to submodules

* Move HDF and Plain Text Geometry operations to submodules

Some changes implemented to the RaGeo and RasGeometry* classes, old methods will be deprecated in 1.0 version coming soon

* Add support for Docker Worker type for RAS Remote execution

* update ai kb